+++ glean/src/core/pings/ping_payload.ts
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-import type { Metrics } from "../metrics/metrics_interface.js";
+import type { Metrics } from "../metrics/index.js";
 import type { JSONObject, JSONArray } from "../utils.js";
 
 export interface PingInfo extends JSONObject {