+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Unreleased changes
 
-[Full changelog](https://github.com/mozilla/glean.js/compare/v0.10.2...main)
+[Full changelog](https://github.com/mozilla/glean.js/compare/v0.11.0...main)
+
+# v0.11.0 (2021-05-03)
+
+[Full changelog](https://github.com/mozilla/glean.js/compare/v0.10.2...v0.11.0)
+
+* [#260](https://github.com/mozilla/glean.js/pull/260): Set minimum node (>= 12.0.0) and npm (>= 7.0.0) versions.
+* [#202](https://github.com/mozilla/glean.js/pull/202): Add a testing API for the ping type.
+* [#253](https://github.com/mozilla/glean.js/pull/253): Implement the timespan metric type.
+* [#261](https://github.com/mozilla/glean.js/pull/261): Show a spinner while setting up python virtual environment
 
 # v0.10.2 (2021-04-26)